chore: Change relative import path for polyCommand
The code was previously using a relative path to import 'polyCommand'. However, it needs to be changed to a more detailed path including index.js for accurate import. Changes are made in cli.js to update the import statement of 'polyCommand' from "./src/commands/poly" to "./src/commands/poly/index.js". This is necessary for the file system to explicitly know which file to import in the given directory.