feat: update src/commands/poly/task.cjs for improved adaptability
Enhanced the robustness and adaptability of src/commands/poly/task.cjs by expanding its functioning with the addition of a new constant 'attributeNames' and changes in the functioning of 'JSXElement' and 'JSXAttribute'. Improved the conditions under which 'translations' are updated and the process of text replacement in the original file.

Also updated the package.json file to pass a 'p' argument to the 'dev' script and added 'verbose: 1' in src/commands/poly/command.js file for detailed output. The latter also incorporates changes to unify translations into a single file post their update. These changes collectively contribute to a more streamlined and user-friendly experience of the 'poly' command.