feat: Add translation variable name to uniteTranslations function
Added a new configuration option "translationVariableName" in the function 'uniteTranslationsIntoOneFile'. This allows to pass the name of the variable holding translations, rather than hardcoding it as "translations". This offers a more flexible solution, allowing various projects to use their own naming convention for the translations variable without the need to modify the source code.